+++ product-service/src/main/java/com/takeit/product/application/dto/category/CategoryResponse.java
@@ -1,4 +1,4 @@
-package com.takeit.product.application.dto;
+package com.takeit.product.application.dto.category;
 
 import java.util.UUID;
