+++ product-service/src/main/java/com/takeit/product/application/dto/product/ProductDetailResponse.java
@@ -1,6 +1,7 @@
-package com.takeit.product.application.dto;
+package com.takeit.product.application.dto.product;
 
 import com.takeit.product.domain.entity.Product;
+
 import java.time.LocalDateTime;
 import java.util.UUID;
