+++ product-service/src/main/java/com/takeit/product/domain/entity/Product.java
@@ -2,16 +2,12 @@ package com.takeit.product.domain.entity;
 
 import com.takeit.common.domain.model.BaseEntity;
 import jakarta.persistence.*;
+import lombok.*;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 @Entity
 @Table(name = "p_product")
