+++ product-service/src/main/java/com/takeit/product/presentation/controller/ProductController.java
@@ -1,42 +1,31 @@
 package com.takeit.product.presentation.controller;
 
 import com.querydsl.core.types.Predicate;
+import com.takeit.common.exception.CustomException;
+import com.takeit.common.exception.ErrorCode;
 import com.takeit.common.presentation.dto.CommonResponse;
-import com.takeit.product.application.dto.ProductDetailResponse;
-import com.takeit.product.application.dto.ProductPageResponse;
-import com.takeit.product.application.dto.ProductResponse;
+import com.takeit.product.application.dto.product.ProductDailyStatResponse;
+import com.takeit.product.application.dto.product.ProductDetailResponse;
+import com.takeit.product.application.dto.product.ProductPageResponse;
+import com.takeit.product.application.dto.product.ProductResponse;
 import com.takeit.product.application.service.ProductService;
 import com.takeit.product.domain.entity.Product;
 import com.takeit.product.presentation.request.CreateProductRequest;
 import com.takeit.product.presentation.request.UpdateProductRequest;
 import jakarta.validation.Valid;
-import java.util.UUID;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
-import java.util.Date;
-
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.querydsl.binding.QuerydslPredicate;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.*;
 
-import org.springframework.data.domain.Page;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.takeit.common.exception.CustomException;
-import com.takeit.common.exception.ErrorCode;
-import com.takeit.product.application.dto.ProductDailyStatResponse;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
+import java.util.Date;
+import java.util.UUID;
 
 @RestController
 @RequiredArgsConstructor
@@ -80,7 +69,7 @@ public class ProductController {
     @GetMapping
     public ResponseEntity<CommonResponse<ProductPageResponse>> getProducts(
             @QuerydslPredicate(root = Product.class) Predicate predicate,
-            @PageableDefault(sort = "id", direction = Sort.Direction.DESC, page = 0, size = 10) Pageable pageable) {
+            @PageableDefault(sort = "id", direction = Sort.Direction.DESC) Pageable pageable) {
         return ResponseEntity.ok().body(CommonResponse.ofSuccess("상품 목록 조회 성공", productService.getProducts(predicate, pageable)));
     }
